Internationalisation des messages v2

+++ src/main/java/com/example/exam/EmployeeService.java
@@ -1,40 +1,50 @@
 package com.example.exam;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Locale;
 
 @Service
 public class EmployeeService {
-        @Autowired
-        private EmployeeRepository employeeRepository;
+    @Autowired
+    private EmployeeRepository employeeRepository;
 
-        public List<Employee> getAllEmployees() {
-            return employeeRepository.findAll();
-        }
+    @Autowired
+    private LocalizationService localizationService;
 
-        public void addEmployee(Employee employee) {
-            employeeRepository.save(employee);
-        }
-
-        public void updateEmployee (Employee employee){
-            Employee oldemployee=employeeRepository.findById(employee.getId());
-
-            oldemployee.setAge(employee.getAge());
-            oldemployee.setEmail(employee.getEmail());
-            oldemployee.setFname(employee.getFname());
-            oldemployee.setLname(employee.getLname());
-            oldemployee.setPhone(employee.getPhone());
-
-            employeeRepository.save(oldemployee);
-        }
-
-        @Transactional
-        public void deleteEmployee (Long id){
-            employeeRepository.deleteById(id);
-
-        }
+    public List<Employee> getAllEmployees() {
+        return employeeRepository.findAll();
     }
 
+    public void addEmployee(Employee employee, Locale locale) {
+
+
+        employeeRepository.save(employee);
+        throw new RuntimeException(localizationService.getMessage("employee.added", locale));
+    }
+
+    public void updateEmployee(Employee employee, Local locale) {
+        Employee existingEmployee = employeeRepository.findById(employee.getId())
+                .orElseThrow(() -> new RuntimeException(localizationService.getMessage("employee.not.found", locale)));
+
+        existingEmployee.setAge(employee.getAge());
+        existingEmployee.setEmail(employee.getEmail());
+        existingEmployee.setFname(employee.getFname());
+        existingEmployee.setLname(employee.getLname());
+        existingEmployee.setPhone(employee.getPhone());
+
+        employeeRepository.save(existingEmployee);
+    }
+
+    @Transactional
+    public void deleteEmployee(Long id, Locale locale) {
+        if (!employeeRepository.existsById(id)) {
+            throw new RuntimeException(localizationService.getMessage("employee.not.found", locale));
+        }
+        employeeRepository.deleteById(id);
+    }
+}
